buzz_data/requirements.txt
google.cloud.storage
requests
beautifulsoup4
